revert guessed artist folder changes
there seems to be a performance issue somewhere
