+++ server/scanner/scanner.go
@@ -217,7 +217,7 @@ func (s *Scanner) populateTrackAndAlbumArtists(tx *db.DB, c *ctx, i int, album *
 	}
 
 	artistName := trags.SomeAlbumArtist()
-	albumArtist, err := s.populateAlbumArtist(tx, album, artistName)
+	albumArtist, err := s.populateAlbumArtist(tx, artistName)
 	if err != nil {
 		return fmt.Errorf("populate artist: %w", err)
 	}
@@ -315,14 +315,11 @@ func populateTrack(tx *db.DB, album *db.Album, albumArtist *db.Artist, track *db
 	return nil
 }
 
-func (s *Scanner) populateAlbumArtist(tx *db.DB, guessedArtistFolder *db.Album, artistName string) (*db.Artist, error) {
+func (s *Scanner) populateAlbumArtist(tx *db.DB, artistName string) (*db.Artist, error) {
 	var artist db.Artist
 	update := db.Artist{
 		Name:     artistName,
 		NameUDec: decoded(artistName),
-	}
-	if guessedArtistFolder != nil {
-		update.GuessedFolderID = guessedArtistFolder.ParentID
 	}
 	if err := tx.Where("name=?", artistName).Assign(update).FirstOrCreate(&artist).Error; err != nil {
 		return nil, fmt.Errorf("find or create artist: %w", err)
@@ -479,8 +476,7 @@ func isCover(name string) bool {
 		"folder.png", "folder.jpg", "folder.jpeg",
 		"album.png", "album.jpg", "album.jpeg",
 		"albumart.png", "albumart.jpg", "albumart.jpeg",
-		"front.png", "front.jpg", "front.jpeg",
-		"artist.png", "artist.jpg", "artist.jpeg":
+		"front.png", "front.jpg", "front.jpeg":
 		return true
 	default:
 		return false
